% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_gr}
\alias{draw_gr}
\title{Plot growth rates}
\usage{
draw_gr(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A plot.
}
\description{
Plot two lines of YoY growth rates of both "export" and "import" over time
}
\examples{
\dontrun{
draw_gr(df)
}
}
