% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_value}
\alias{draw_value}
\title{Plot volume}
\usage{
draw_value(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A plot.
}
\description{
Plot two lines of volumes of both "export" and "import" over time
}
\examples{
\dontrun{
draw_value(df)
}
}
