% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{show_table}
\alias{show_table}
\title{Show table}
\usage{
show_table(df, em, ar_go)
}
\arguments{
\item{df}{A dataframe.}

\item{em}{either "export" or "import"}

\item{ar_go}{either area or goods}
}
\value{
A DT table
}
\description{
Show table in DT format
}
\examples{
\dontrun{
show_table(df, em = "export", ar_go = area)
}
}
