% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade}
\alias{trade}
\title{Trade data by month}
\format{
A dataframe.
\describe{
\item{\verb{Exp or Imp}}{"export" or "import"}
\item{Area}{areas}
\item{goods}{goods}
\item{month}{dates}
\item{value}{value}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
trade
}
\description{
Trade data by month.
}
\keyword{datasets}
