% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_year}
\alias{trade_year}
\title{Trade data by year}
\format{
A dataframe.
\describe{
\item{\verb{Exp or Imp}}{"export" or "import"}
\item{Area}{areas}
\item{goods}{goods}
\item{year}{year}
\item{value}{value}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
trade_year
}
\description{
Trade data by year.
}
\keyword{datasets}
