% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{levels_area}
\alias{levels_area}
\title{Areas}
\format{
A character vector
\describe{
\item{distinct areas}{"China", "EU", R Korea", "USA"}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
levels_area
}
\description{
Areas by alphabetical order.
}
\keyword{datasets}
