% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{levels_goods}
\alias{levels_goods}
\title{Goods}
\format{
A character vector.
\describe{
\item{goods}{IC}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
levels_goods
}
\description{
Goods.
}
\keyword{datasets}
