% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade}
\alias{trade}
\title{Japan trade data by month}
\format{
Dataframe
\describe{
\item{\verb{Exp or Imp}}{"export" or "import"}
\item{Area}{Area category}
\item{month}{date}
\item{goods}{goods category}
\item{value}{billion yen per month}
\item{gr}{YoY growth rates}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
trade
}
\description{
Monthly data from 2009.
}
\keyword{datasets}
