% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_year}
\alias{trade_year}
\title{Japan trade data by year}
\format{
Dataframe
\describe{
\item{\verb{Exp or Imp}}{factor of "export" or "import"}
\item{Area}{Area category}
\item{goods}{goods category}
\item{year}{double}
\item{value}{billion yen per year}
}
}
\source{
http://www.customs.go.jp/toukei/info/index_e.htm
}
\usage{
trade_year
}
\description{
Annual data from 2009.
}
\keyword{datasets}
